renamed file to listTest, added another test for an empty list, added another layer of description
